# src/app/cryptoperuser/controller.js
const httpStatus = require('http-status');
const cryptoPerUser = require('./criptoperuser');

exports.addCoinForFollowUp = async (req, res) => {
    try {
        const response = await cryptoPerUser.addCoinForFollowUp(req.user, req.body.coin);

        if (!response) return res.status(httpStatus.NOT_FOUND).send({ message: 'Unable to add coin for follow up' })

        return res.status(httpStatus.OK).send({ message: "Coin added" })
    } catch (error) {
        return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: "Internal server error", error: error.message })
    }
}

exports.getTopCrypto = async (req, res) => {
    try {
        query = req.query.top;
        const response = await cryptoPerUser.topCryptos(req.user, req.query.limit);

        if (!response) return res.status(httpStatus.NOT_FOUND).send({ message: 'Unable to get your top crypto coins' })

        return res.status(httpStatus.OK).send({ message: "Your top crypto", data: response })
    } catch (error) {
        return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: "Internal server error", error: error.message })
    }

}

exports.getAllCoins = async (req, res) => {
    try {
        const response = await cryptoPerUser.getAllCoins(req.user, req.coin);

        if (!response) return res.status(httpStatus.NOT_FOUND).send({ message: 'Unable to get all coins' })

        return res.status(httpStatus.OK).send({ message: "allCoins", data: response })
    } catch (error) {
        return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: "Internal server error", error: error.message })
    }

}